Fixes to transaction and labels

+++ team-18-web-project/app/controllers/ActionController.java
@@ -63,7 +63,7 @@ public class ActionController extends Controller {
 
     /**
      * Confirms transaction
-     * @return HTTP response to create sale request
+     * @return HTTP response to confirm transaction request
      */
     @Authenticated(Secured.class)
     public Result confirmTransaction(int saleId, int tranId) {
@@ -71,53 +71,53 @@ public class ActionController extends Controller {
         User user = User.findByUsername(session("username"));
         DynamicForm f = Form.form().bindFromRequest();
 
-        // Handle transaction cancel form
-        if (f.get("cancelTransactionId") != null) {
-            int transactionId;
-            try { // Convert string to integer
-                transactionId = Integer.parseInt(f.get("cancelTransactionId"));
-            } catch (NumberFormatException e) {
-                e.printStackTrace();
-                return notFound404();
+        if (user != null && user.canBeSeller(saleId)) {
+            // Handle transaction cancel form
+            if (f.get("cancelTransactionId") != null) {
+                int transactionId;
+                try { // Convert string to integer
+                    transactionId = Integer.parseInt(f.get("cancelTransactionId"));
+                } catch (NumberFormatException e) {
+                    e.printStackTrace();
+                    return notFound404();
+                }
+
+                // Delete line items
+                List<LineItem> list = LineItem.findByTransactionId(tranId);
+                for (LineItem li : list) {
+                    li.delete();
+                }
+                return redirect("/sale/" + saleId);
             }
 
-            // Delete line items
-            List<LineItem> list = LineItem.findByTransactionId(tranId);
-            for (LineItem li : list) {
-                li.delete();
-            }
-            return redirect("/sale/" + saleId);
-        }
+            //enters a confirmed transaction updating the buyer info using a form
+            if (f.get("confirmTransaction") != null) {
+                Sale s = Sale.findById(saleId);
+                Transaction t = Transaction.findById(tranId);
+                t.setBuyerName(f.get("name"));
+                t.setBuyerAddress(f.get("address"));
+                t.setBuyerEmail(f.get("email"));
+                t.setCompleted(1);
+                t.setPaymentMethod(f.get("payment"));
+                t.save();
 
-        //enters a confirmed transaction updating the buyer info using a form
-        if (f.get("confirmTransaction") != null) {
+                //update item quantity for sale items
+                List<SaleItem> saleItems = s.getItems();
+                List<LineItem> lineItems = t.getLineItems();
+                for (LineItem li : lineItems) {
+                    for (SaleItem si: saleItems) {
 
-            Sale s = Sale.findById(saleId);
-            Transaction t = Transaction.findById(tranId);
-            t.setBuyerName(f.get("name"));
-            t.setBuyerAddress(f.get("address"));
-            t.setBuyerEmail(f.get("email"));
-            t.setCompleted(1);
-            t.setPaymentMethod(f.get("payment"));
-            t.save();
+                        if (li.getId() == si.getId()) {
 
-            //update item quantity for sale items
-            List<SaleItem> saleItems = s.getItems();
-            List<LineItem> lineItems = t.getLineItems();
-            for (LineItem li : lineItems) {
-                for (SaleItem si: saleItems) {
-
-                    if (li.getId() == si.getId()) {
-
-                        si.setQuantity(si.getQuantity() - li.getQuantity());
-                        si.save();
+                            si.setQuantity(si.getQuantity() - li.getQuantity());
+                            si.save();
+                        }
                     }
                 }
             }
-
+            return redirect("/sale/" + saleId + "/transactionReceipt/" + tranId);
         }
-
-        return redirect("/sale/" + saleId + "/transactionReceipt/" + tranId);
+        return notFound404();
     }
     /**
      * Creates a sale
@@ -425,7 +425,7 @@ public class ActionController extends Controller {
         User u = User.findByUsername(session("username"));
         Transaction t = Transaction.findById(tranId);
         if (s != null && u != null && u.canBeSeller(saleId) && t != null) {
-                // Check if sale exists, user exists and can be a seller, and transaction exists
+            // Check if sale exists, user exists and can be a seller, and transaction exists
 
             DynamicForm f = Form.form().bindFromRequest();
 
@@ -551,7 +551,10 @@ public class ActionController extends Controller {
     }
 
 
-
+    /**
+     * Upload a profile picture (not working yet)
+     * @return HTTP response to upload profile picture request
+     */
     @Authenticated(Secured.class)
     public Result uploadProfilePicture() {
         MultipartFormData body = request().body().asMultipartFormData();
